test(ci): checkout-less E2E across npm/pnpm/bun, online + offline, all reports
Run the preview E2E without checking out the repo so each package runner uses an
empty working directory (a true end-user environment) instead of inheriting the
repo's pnpm-workspace.yaml / package.json — which is why pnpm previously needed
--config.block-exotic-subdeps / --config.minimum-release-age. Those overrides
are gone.

- The fixture is built in the publish job (which has a checkout) and uploaded as
  an artifact; the preview URL is exposed as a job output. E2E jobs download the
  fixture, point CLAUDE_CONFIG_DIR at it, and run from an empty directory.
- One unified run step switches on the runner (npx / pnpm dlx / bunx -p) and
  exercises daily, monthly, and session — both offline and online (online
  fetches live pricing) — validating the JSON output each time.
- pnpm dlx on Windows is skipped (it can't install the optional native binary).

e2e-setup.mjs is replaced by generate-e2e-fixture.mjs (fixture only); URL
resolution moved into the publish job.
