+++ .github/workflows/ci.yaml
@@ -144,6 +144,7 @@ jobs:
     runs-on: blacksmith-32vcpu-ubuntu-2404-arm
     outputs:
       urls: ${{ steps.publish.outputs.urls }}
+      preview-url: ${{ steps.preview-url.outputs.url }}
 
     steps:
       - uses: actions/checkout@df4cb1c069e1874edd31b4311f1884172cec0e10 # v6.0.3
@@ -166,6 +167,20 @@ jobs:
           done
       - id: publish
         run: pnpm pkg-pr-new publish --pnpm --bin './apps/ccusage' './packages/ccusage-linux-arm64' './packages/ccusage-linux-x64' './packages/ccusage-darwin-arm64' './packages/ccusage-darwin-x64' './packages/ccusage-win32-x64'
+      # Resolve the ccusage CLI preview URL and build the E2E fixture here (this
+      # job has the checkout); the E2E jobs then run without a checkout.
+      - id: preview-url
+        env:
+          URLS: ${{ steps.publish.outputs.urls }}
+        run: |
+          node -e 'const urls=(process.env.URLS??"").split(/\s+/).filter(Boolean);const u=urls.find((x)=>/(?:^|\/)ccusage@/.test(x));if(u==null){throw new Error(`ccusage preview URL not found in: ${urls.join(", ")}`)}require("node:fs").appendFileSync(process.env.GITHUB_OUTPUT,`url=${u}\n`)'
+      - name: Generate E2E fixture
+        run: node .github/scripts/generate-e2e-fixture.mjs "$RUNNER_TEMP/e2e-fixture"
+      - uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
+        with:
+          name: ccusage-e2e-fixture
+          path: ${{ runner.temp }}/e2e-fixture
+          if-no-files-found: error
 
   ccusage-preview-e2e:
     # Run on PRs and on pushes to the default branch. `push` only fires on main
@@ -178,6 +193,14 @@ jobs:
       - npm-publish-dry-run-and-upload-pkg-pr-now
     runs-on: ${{ matrix.os }}
     timeout-minutes: 20
+    # No checkout on purpose: the working directory stays empty so each package
+    # runner behaves like a real end user instead of inheriting the repo's
+    # pnpm-workspace.yaml / package.json (which is why pnpm previously needed
+    # --config overrides). The preview URL and the fixture come from the publish
+    # job via a job output and an artifact.
+    env:
+      CCUSAGE_PREVIEW_URL: ${{ needs.npm-publish-dry-run-and-upload-pkg-pr-now.outputs.preview-url }}
+      LOG_LEVEL: '0'
     strategy:
       fail-fast: false
       matrix:
@@ -200,49 +223,66 @@ jobs:
           - bun
 
     steps:
-      - uses: actions/checkout@df4cb1c069e1874edd31b4311f1884172cec0e10 # v6.0.3
       - uses: actions/setup-node@48b55a011bda9f5d6aeb4c2d9c7362e8dae4041e # v6.4.0
         with:
           node-version: lts/*
       - uses: pnpm/action-setup@b906affcce14559ad1aafd4ab0e942779e9f58b1 # v4
         if: matrix.runner == 'pnpm'
+        with:
+          version: latest
       - uses: oven-sh/setup-bun@0c5077e51419868618aeaa5fe8019c62421857d6 # v2
         if: matrix.runner == 'bun'
-      - name: Prepare preview package E2E
+        with:
+          bun-version: latest
+      - uses: actions/download-artifact@3e5f45b2cfb9172054b4087a40e8e0b5a5461e7c # v8.0.1
+        with:
+          name: ccusage-e2e-fixture
+          path: ${{ runner.temp }}/e2e-fixture
+      - name: Point CLAUDE_CONFIG_DIR at the fixture
+        shell: pwsh
+        run: |
+          "CLAUDE_CONFIG_DIR=$env:RUNNER_TEMP/e2e-fixture" >> $env:GITHUB_ENV
+      - name: Run preview E2E
+        shell: pwsh
         env:
-          PKG_PR_NEW_URLS: ${{ needs.npm-publish-dry-run-and-upload-pkg-pr-now.outputs.urls }}
-        run: node .github/scripts/e2e-setup.mjs
-      - name: Run preview E2E with npx
-        if: matrix.runner == 'npm'
-        shell: pwsh
+          RUNNER_KIND: ${{ matrix.runner }}
+          # pnpm 11 defaults minimumReleaseAge to 1440 (24h), which would reject
+          # a pkg.pr.new preview published moments ago. Opt out via the config
+          # env var (cleaner than a --config flag, and only pnpm reads it).
+          # https://pnpm.io/settings#minimumreleaseage
+          npm_config_minimum_release_age: '0'
         run: |
-          npx --yes "$env:CCUSAGE_PREVIEW_URL" --version
-          npx --yes "$env:CCUSAGE_PREVIEW_URL" daily --offline --json | node -e "JSON.parse(require('fs').readFileSync(0, 'utf8'))"
-          npx --yes "$env:CCUSAGE_PREVIEW_URL" daily --offline
-      - name: Run preview E2E with pnpm dlx
-        # pnpm dlx on Windows doesn't install optional dependency native
-        # binaries (ccusage.exe) into the temporary virtual store, causing
-        # the CLI launcher's spawn to fail with ENOENT.
-        if: matrix.runner == 'pnpm' && runner.os != 'Windows'
-        shell: pwsh
-        # On Windows, `--config.package-import-method=copy` works around a
-        # pnpm bug where the dlx store hardlinks the platform-native optional
-        # dependency package directory but leaves `bin/ccusage.exe` missing,
-        # causing the launcher to fail with `spawn ENOENT`. Non-Windows runners
-        # keep pnpm's default (hardlink) for faster dlx setup.
-        run: |
-          $pnpmArgs = @('--config.block-exotic-subdeps=false', '--config.minimum-release-age=0')
-          if ($IsWindows) { $pnpmArgs += '--config.package-import-method=copy' }
-          pnpm @pnpmArgs dlx "$env:CCUSAGE_PREVIEW_URL" --version
-          pnpm @pnpmArgs dlx "$env:CCUSAGE_PREVIEW_URL" daily --offline --json | node -e "JSON.parse(require('fs').readFileSync(0, 'utf8'))"
-          pnpm @pnpmArgs dlx "$env:CCUSAGE_PREVIEW_URL" daily --offline
-      - name: Run preview E2E with bunx
-        if: matrix.runner == 'bun'
-        shell: pwsh
-        run: |
-          bunx -p "$env:CCUSAGE_PREVIEW_URL" ccusage --version
-          bunx -p "$env:CCUSAGE_PREVIEW_URL" ccusage daily --offline --json | node -e "JSON.parse(require('fs').readFileSync(0, 'utf8'))"
-          bunx -p "$env:CCUSAGE_PREVIEW_URL" ccusage daily --offline
+          # pnpm dlx on Windows doesn't install optional native binaries into its
+          # temporary store, so the CLI launcher fails with spawn ENOENT; skip
+          # that one combination.
+          if ($env:RUNNER_KIND -eq 'pnpm' -and $IsWindows) {
+            Write-Host 'skip: pnpm dlx on Windows cannot install the native binary'
+            exit 0
+          }
+          $url = $env:CCUSAGE_PREVIEW_URL
+          function Invoke-Ccusage([string[]]$ccArgs) {
+            switch ($env:RUNNER_KIND) {
+              'npm' { return (npx --yes $url @ccArgs) }
+              'pnpm' { return (pnpm dlx $url @ccArgs) }
+              'bun' { return (bunx -p $url ccusage @ccArgs) }
+            }
+          }
+          Invoke-Ccusage @('--version') | Out-Host
+          if ($LASTEXITCODE -ne 0) { throw '--version failed' }
+          # Exercise every report command, both offline and online (online fetches
+          # live pricing), validating the JSON output each time.
+          foreach ($cmd in 'daily', 'monthly', 'session') {
+            foreach ($mode in @(@('--offline'), @())) {
+              $label = "$cmd $($mode -join ' ')".Trim()
+              Invoke-Ccusage (@($cmd) + $mode) | Out-Host
+              if ($LASTEXITCODE -ne 0) { throw "ccusage $label failed ($LASTEXITCODE)" }
+              $json = Invoke-Ccusage (@($cmd) + $mode + @('--json'))
+              if ($LASTEXITCODE -ne 0) { throw "ccusage $label --json failed ($LASTEXITCODE)" }
+              $json | node -e "JSON.parse(require('fs').readFileSync(0, 'utf8'))"
+              if ($LASTEXITCODE -ne 0) { throw "ccusage $label --json produced invalid JSON" }
+              Write-Host "ok: $label"
+            }
+          }
 
   ccusage-perf-comment:
     name: ccusage perf comment
